% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_dates.R
\name{utils-dates}
\alias{utils-dates}
\alias{end_of_month}
\alias{start_of_month}
\title{Date Utility Functions}
\usage{
end_of_month(date)

start_of_month(date)
}
\arguments{
\item{date}{Input date (\code{POSIXct} or \code{character} string)}
}
\value{
Returns a \code{Date} object (i.e. \code{POSIXct})

Returns a \code{Date} object (i.e. \code{POSIXct})
}
\description{
Date utility functions and helpers for manipulating and managing dates.

Retrieves the last day of the provided date's current month.

Retrieves the first day of the provided date's current month.
}
\examples{
# end_of_month ------------------------------------------------------------

# character input
end_of_month("2020-08-13")

# date input
end_of_month(as.Date("2020-08-13"))
# start_of_month ----------------------------------------------------------

# character input
start_of_month("2020-08-13")

# date input
start_of_month(as.Date("2020-08-13"))
}
\seealso{
\code{\link[lubridate:round_date]{lubridate::ceiling_date()}}

\code{\link[lubridate:round_date]{lubridate::floor_date()}}
}
