% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyreserving-package.R
\docType{package}
\name{shinyreserving-package}
\alias{shinyreserving}
\alias{shinyreserving-package}
\title{shinyreserving: Actuarial Reserving}
\description{
Actuarial Loss Reserving and Development R Package.
}
\author{
\strong{Maintainer}: Jimmy Briggs \email{jimmy.briggs@jimbrig.com} (\href{https://orcid.org/0000-0002-7489-8787}{ORCID})

}
\keyword{internal}
